Create new Marxan Input Parameter File (input.dat). See the [Marxan User Manual](https://marxansolutions.org/wp-content/uploads/2021/02/Marxan-User-Manual_2021.pdf) for more information: Serra, N., Kockel, A., Game, E. T., Grantham H., Possingham H.P., & McGowan, J. (2020). Marxan User Manual: For Marxan version 2.43 and above. The Nature Conservancy (TNC), Arlington, Virginia, United States and Pacific Marine Analysis and Research Association (PacMARA), Victoria, British Columbia, Canada.

Create new Marxan Input Parameter File (input.dat). See the [Marxan User Manual](https://marxansolutions.org/wp-content/uploads/2021/02/Marxan-User-Manual_2021.pdf) for more information: Serra, N., Kockel, A., Game, E. T., Grantham H., Possingham H.P., & McGowan, J. (2020). Marxan User Manual: For Marxan version 2.43 and above. The Nature Conservancy (TNC), Arlington, Virginia, United States and Pacific Marine Analysis and Research Association (PacMARA), Victoria, British Columbia, Canada.

Usage

newParams(
  BLM = 1,
  PROP = 0.5,
  RANDSEED = -1,
  NUMREPS = 10L,
  NUMITNS = 1000000L,
  STARTTEMP = -1,
  NUMTEMP = 10000,
  COSTTHRESH = 0,
  THRESHPEN1 = 0,
  THRESHPEN2 = 0,
  INPUTDIR = "input",
  PUNAME = "pu.dat",
  SPECNAME = "spec.dat",
  PUVSPRNAME = "puvspr.dat",
  BOUNDNAME = "bound.dat",
  SCENNAME = "output",
  SAVERUN = 3L,
  SAVEBEST = 3L,
  SAVESUMMARY = 3L,
  SAVESCEN = 3L,
  SAVETARGMET = 3L,
  SAVESUMSOLN = 3L,
  SAVEPENALTY = 3L,
  SAVELOG = 2L,
  OUTPUTDIR = "output",
  RUNMODE = 1L,
  MISSLEVEL = 1,
  ITIMPTYPE = 0L,
  HEURTYPE = -1L,
  CLUMPTYPE = 0L,
  VERBOSITY = 3L,
  SAVESOLUTIONSMATRIX = 3L
)

Arguments

BLM

A numeric value for the Boundary Length Modifier. Default is 1

PROP

Proportion of planning units in initial reserve system. Default is 0.5

RANDSEED

Random seed number. Default is -1

NUMREPS

Number (integer) of repeat runs (or solutions). Default is 10L

NUMITNS

Number (integer) of iterations for annealing. Default is 1000000L

STARTTEMP

A numeric value for the starting temperature for annealing. Default is -1

NUMTEMP

A numeric value for the temperature decreases for annealing. Default is 10000

COSTTHRESH

A numeric value for the cost threshold. Default is 0

THRESHPEN1

A numeric value for the size of cost threshold penalty. Default is 0

THRESHPEN2

A numeric value for the shape of cost threshold penalty. Default is 0

INPUTDIR

A character string naming the folder containing input data files. Default is "input".

PUNAME

A character string naming the Planning Unit File. Default is "pu.dat".

SPECNAME

A character string naming the Conservation Feature File. Default is "spec.dat".

PUVSPRNAME

A character string naming the Planning Unit versus Conservation Feature File. Default is "puvspr.dat".

BOUNDNAME

A character string naming the Boundary Length File. Default is "bound.dat".

SCENNAME

A character string for a scenario name appended to all the saved output files. Default is "output".

SAVERUN

An integer code to save each run. A value of 0L will not save a file, 1L saves the file as .dat, 2L saves the file as .txt, and 3L (default) saves the file as .csv.

SAVEBEST

An integer code to save the best run. A value of 0L will not save a file, 1L saves the file as .dat, 2L saves the file as .txt, and 3L (default) saves the file as .csv.

SAVESUMMARY

An integer code to save summary information. A value of 0L will not save a file, 1L saves the file as .dat, 2L saves the file as .txt, and 3L (default) saves the file as .csv.

SAVESCEN

An integer code to save scenario information. A value of 0L will not save a file, 1L saves the file as .dat, 2L saves the file as .txt, and 3L (default) saves the file as .csv.

SAVETARGMET

An integer code to save targets met information. A value of 0L will not save a file, 1L saves the file as .dat, 2L saves the file as .txt, and 3L (default) saves the file as .csv.

SAVESUMSOLN

An integer code to save summed solution information. A value of 0L will not save a file, 1L saves the file as .dat, 2L saves the file as .txt, and 3L (default) saves the file as .csv.

SAVEPENALTY

An integer code to save computed feature penalties. A value of 0L will not save a file, 1L saves the file as .dat, 2L saves the file as .txt, and 3L (default) saves the file as .csv.

SAVELOG

An integer code to save log files. A value of 0L will not save a file, 1L saves the file as .dat, 2L (default) saves the file as .txt, and 3L saves the file as .csv.

OUTPUTDIR

A character string naming the folder in which to save output files. Default is "output".

RUNMODE

An Integer code that sets the run option. 0L Apply Simulated Annealing followed by a Heuristic; 1L (default) Apply Simulated Annealing followed by Iterative Improvement; 2L Apply Simulated Annealing followed by a Heuristic, followed by Iterative Improvement; 3L Use only a Heuristic; 4L Use only Iterative Improvement; 5L Use a Heuristic followed by Iterative Improvement; 6L Use only Simulated Annealing

MISSLEVEL

A numeric proportion representing the proportion of the target a conservation feature must reach in order for it to be reported as met. Default value is 1 (i.e. 100 %)

ITIMPTYPE

An integer code that sets the iterative improvement. 0L (default) Normal Iterative Improvement; 1L Two Step Iterative Improvement; 2L ‘Swap’ Iterative Improvement; 3L Normal Improvement followed by Two Step Iterative Improvement

HEURTYPE

An integer code that sets the heuristic.0L Richness; 1L Greedy; 2L Max Rarity; 3L Best Rarity; 4L Average Rarity; 5L Sum Rarity; 6L Product Irreplaceability; 7L Summation Irreplaceability

CLUMPTYPE

An integer code that sets the clumping rule. A value of 0L (default) is Partial clumps do not count– Clumps smaller than the target score nothing, 1L is Partial clumps count half– Clumps smaller than the target score half their amount, and 2L Graduated penalty– Score is proportional to the size of the clump

VERBOSITY

An integer code that sets screen printing options. A value of 0L is for Silent Running – Only the title of the program is displayed, 1L is Results Only – Marxan will display which run it is up to, the basic results of each run and the total run time, 2L is General Progress – In addition to the information about each run, Marxan will display information on the data that has been read in as well as details on any conservation features whose targets and requirements are such that they cannot be adequately reserved in the system, and 3L (default) is Detailed Progress – Shows exactly where the program is up to and gives the value of the system each time the temperature changes.

SAVESOLUTIONSMATRIX

An integer code to save planning units selected in for each run. A value of 0L will not save a file, 1L saves the file as .dat, 2L saves the file as .txt, and 3L (default) saves the file as .csv.

Value

A 'marxanParams' class object

Examples

# Accepting all default values
inputdat <- newParams()

# Accepting all default values except BLM
inputdat <- newParams(BLM = 1000)